cleaning: move contract file in helpers folder
